refactor: share iOS XML parsing helpers
